requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
asyncpg==0.31.0
python-jose==3.5.0
bcrypt==4.1.3
passlib==1.7.4
pydantic==2.12.5
pydantic-settings==2.9.1
python-multipart==0.0.20
